Implemented authentication headers
