+++ src/screens/ScanPassportScreen.tsx
@@ -81,14 +81,15 @@ function ScanPassportScreen({navigation}: Props) {
   }, [error, uiState, modalIsVisible])
 
   // test a working/not working code in the simulator
-  // useEffect(() => {
-  //   const good = '86d89f24-fc11-4829-aa4e-5daee20a370a'
-  //   const bad = 'fdsfds'
-  //   onBarCodeRead({
-  //     data: good,
-  //     type: RNCamera.Constants.BarCodeType.qr,
-  //   })
-  // }, [])
+  /*
+  useEffect(() => {
+    const good = '86d89f24-fc11-4829-aa4e-5daee20a370a'
+    const bad = 'fdsfds'
+    onBarCodeRead({
+      data: good,
+      type: RNCamera.Constants.BarCodeType.qr,
+    })
+  }, []) */
 
   return (
     <>
